Move federation into 'impl'.
