+++ impl/src/main/scala/quasar/impl/evaluate/Source.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.evaluate
+package quasar.impl.evaluate
 
 import quasar.common.resource.ResourcePath
