+++ run/src/main/scala/quasar/run/ResourceRouter.scala
@@ -21,8 +21,8 @@ import quasar.common.resource._
 import quasar.contrib.cats.effect._
 import quasar.contrib.pathy.AFile
 import quasar.contrib.std.uuid._
-import quasar.evaluate.Source
 import quasar.impl.datasources.DatasourceManagement.Running
+import quasar.impl.evaluate.Source
 import quasar.mimir.evaluate.QueryAssociate
 import quasar.run.optics.{stringUUIDP => UuidString}
